Refactor: Command/Query 패키지 구조 분리 및 어노테이션 순서 통일
