+++ src/main/java/team/unibusk/backend/domain/performance/presentation/query/PerformanceQueryDocsController.java
@@ -1,4 +1,4 @@
-package team.unibusk.backend.domain.performance.presentation;
+package team.unibusk.backend.domain.performance.presentation.query;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
